Add test for GeoHashUtils#bbox()
