+++ core/src/test/java/org/elasticsearch/common/geo/GeoHashTests.java
@@ -18,6 +18,7 @@
  */
 package org.elasticsearch.common.geo;
 
+import org.apache.lucene.geo.Rectangle;
 import org.elasticsearch.test.ESTestCase;
 
 /**
@@ -56,5 +57,17 @@ public class GeoHashTests extends ESTestCase {
                 }
             }
         }
+    }
+
+    public void testBboxFromHash() {
+        String hash = randomGeohash(1, 12);
+        int level = hash.length();
+        Rectangle bbox = GeoHashUtils.bbox(hash);
+        // check that the length is as expected
+        double expectedLonDiff = 360.0 / (Math.pow(8.0, (level + 1) / 2) * Math.pow(4.0, level / 2));
+        double expectedLatDiff = 180.0 / (Math.pow(4.0, (level + 1) / 2) * Math.pow(8.0, level / 2));
+        assertEquals(expectedLonDiff, bbox.maxLon - bbox.minLon, 0.00001);
+        assertEquals(expectedLatDiff, bbox.maxLat - bbox.minLat, 0.00001);
+        assertEquals(hash, GeoHashUtils.stringEncode(bbox.minLon, bbox.minLat, level));
     }
 }
